+++ apps/frontend/src/containers/TicketRecent/TicketRecent.types.ts
@@ -1,0 +1,9 @@
+import { Components } from '@o2s/api-harmonization';
+
+export interface TicketRecentProps {
+    id: string;
+    accessToken: string;
+    locale: string;
+}
+
+export type TicketRecentPureProps = TicketRecentProps & Components.TicketRecent.Model.TicketRecentComponent;