+++ packages/framework/src/modules/cms/models/components/tickets-recent.model.ts
@@ -1,0 +1,14 @@
+import { Component } from '@/utils/models';
+
+export class TicketRecentComponent extends Component.Component {
+    title?: string;
+    noResults?: string;
+    commentsTitle?: string;
+    labels!: {
+        today: string;
+        yesterday: string;
+        details: string;
+    };
+    limit!: number;
+    detailsUrl!: string;
+}