+++ packages/integrations/mocked/src/utils/delay.ts
@@ -1,4 +1,7 @@
 import { delay } from 'rxjs';
 
 // fake delay on mocked responses to simulate network delays
-export const responseDelay = <T>() => delay<T>(Math.random() * 200 + 100);
+export const responseDelay = <T>() => {
+    const isEnabled = process.env.MOCKED_INTEGRATION_DELAYS_ENABLED === 'true';
+    return delay<T>(isEnabled ? Math.random() * 200 + 100 : 0);
+};