+++ apps/api-harmonization/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @o2s/api-harmonization
 
+## 0.12.0
+
+### Minor Changes
+
+- b4cddfb: add seo, add headers
+
+### Patch Changes
+
+- f2a6781: fixed an issue with alternative URLs for pages - on pages with dynamic URLs (e.g. /cases/(.+)) switching to another locale caused route to change to /cases/(.+) instead of /cases/12345
+- eea2896: added recent tickets component
+- Updated dependencies [f2a6781]
+- Updated dependencies [eea2896]
+- Updated dependencies [b4cddfb]
+- Updated dependencies [54c9fb5]
+    - @o2s/integrations.strapi-cms@0.11.0
+    - @o2s/integrations.mocked@0.11.0
+    - @o2s/framework@0.12.0
+
 ## 0.11.0
 
 ### Minor Changes